refactor: move api creation to Dependencies and Add ApiFactory
