+++ shared/src/commonMain/kotlin/com/akjaw/ai/assistant/shared/chat/domain/ChatMessageHandler.kt
@@ -4,33 +4,20 @@ import app.cash.sqldelight.coroutines.asFlow
 import app.cash.sqldelight.coroutines.mapToList
 import com.akjaw.ai.assistant.database.Database
 import com.akjaw.ai.assistant.shared.Endpoints
+import com.akjaw.ai.assistant.shared.chat.data.api.ApiFactory
 import com.akjaw.ai.assistant.shared.chat.domain.model.ChatMessage
 import com.akjaw.ai.assistant.shared.dashboard.domain.ChatType
 import io.ktor.client.HttpClient
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.map
-import kotlinx.serialization.json.Json
-
-private fun createAddTickTickTask(client: HttpClient) = AddTask(
-    client = client,
-    endpointUrl = Endpoints.AddTaskTickTick.URL,
-    auth = Endpoints.AddTaskTickTick.AUTH
-)
-
-private fun createAddNotionTask(client: HttpClient) = AddTask(
-    client = client,
-    endpointUrl = Endpoints.AddTaskNotion.URL,
-    auth = Endpoints.AddTaskNotion.AUTH
-)
 
 class ChatMessageHandler(
-    client: HttpClient = createKtorClient(),
+    apiFactory: ApiFactory,
     private val database: Database
 ) {
-    private val addNotionTask = createAddNotionTask(client)
-    private val addTickTickTask = createAddTickTickTask(client)
+    private val addNotionTask = apiFactory.createAddNotionTask()
+    private val addTickTickTask = apiFactory.createAddTickTickTask()
 
     fun getMessagesForType(chatType: ChatType): Flow<List<ChatMessage>> =
         database.messageEntityQueries.selectByType(chatType)
