+++ shared/src/commonMain/kotlin/com/akjaw/ai/assistant/shared/composition/Dependencies.kt
@@ -1,14 +1,66 @@
 package com.akjaw.ai.assistant.shared.composition
 
 import com.akjaw.ai.assistant.database.Database
+import com.akjaw.ai.assistant.shared.chat.data.api.ApiFactory
+import com.akjaw.ai.assistant.shared.chat.data.api.ProductionApiFactory
 import com.akjaw.ai.assistant.shared.chat.domain.ChatMessageHandler
-import com.akjaw.ai.assistant.shared.chat.domain.createKtorClient
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.plugins.HttpTimeout
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.plugins.defaultRequest
+import io.ktor.client.plugins.logging.LogLevel
+import io.ktor.client.plugins.logging.Logger
+import io.ktor.client.plugins.logging.Logging
+import io.ktor.http.ContentType
+import io.ktor.http.contentType
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.json.Json
 
 object Dependencies {
 
     internal lateinit var database: Database
 
+    internal val jsonSerialization = Json {
+        prettyPrint = true
+        isLenient = true
+    }
+
+    private val apiFactory: ApiFactory by lazy {
+        ProductionApiFactory(createKtorClient())
+    }
+
     internal val chatMessageHandler: ChatMessageHandler by lazy {
-        ChatMessageHandler(createKtorClient(), database)
+        ChatMessageHandler(apiFactory, database)
+    }
+
+    internal fun createKtorClient(engine: HttpClientEngine? = null): HttpClient {
+        val config: HttpClientConfig<*>.() -> Unit = {
+            install(Logging) {
+                level = LogLevel.ALL
+                logger = object : Logger {
+                    override fun log(message: String) {
+                        co.touchlab.kermit.Logger.i(tag = "Ktor") { message }
+                    }
+                }
+            }
+            install(ContentNegotiation) {
+                json()
+            }
+            install(HttpTimeout) {
+                requestTimeoutMillis = 30_000
+                connectTimeoutMillis = 30_000
+                socketTimeoutMillis = 30_000
+            }
+            defaultRequest {
+                contentType(ContentType.Application.Json)
+            }
+        }
+        return if (engine == null) {
+            HttpClient(config)
+        } else {
+            HttpClient(engine, config)
+        }
     }
 }
